Commit fronts
Commit fronts

+++ src/pages/resources/Index.jsx
@@ -7,7 +7,8 @@ import {
 import { Link } from "react-router-dom";
 import { routesProtection } from "../../assets/routesProtection";
 import { useNavigate } from "react-router-dom";
-import ResourceCreate from "../resources/Create"; 
+import ResourceCreate from "../resources/Create";
+import Swal from "sweetalert2";
 
 const ResourceIndex = () => {
   const { resources, setResources, roleInSession, showNotification } =
@@ -148,14 +149,14 @@ const ResourceIndex = () => {
       }
     });
 
-    const Navbar = () => {
-      const [showPersonal, setShowPersonal] = useState(true);
-      const [selectedTable, setSelectedTable] = useState("materiales");
-    
-      const hidePersonalOption = () => {
-        setShowPersonal(false);
-      };
-    }
+  const Navbar = () => {
+    const [showPersonal, setShowPersonal] = useState(true);
+    const [selectedTable, setSelectedTable] = useState("materiales");
+
+    const hidePersonalOption = () => {
+      setShowPersonal(false);
+    };
+  };
   // Obtener los recursos para la página actual
   const startIndex = (currentPage - 1) * itemsPerPage;
   const endIndex = startIndex + itemsPerPage;
@@ -166,12 +167,10 @@ const ResourceIndex = () => {
   const currentResourcesp = filteredResources.slice(startIndexp, endIndexp);
 
   const [showCreateForm, setShowCreateForm] = useState(false);
-  
 
   const toggleResourcesCreate = () => {
     setShowCreateForm(!showCreateForm);
   };
-
 
   return (
     <div className="flex-container">
@@ -203,16 +202,15 @@ const ResourceIndex = () => {
             <h1 className="text-2xl font-semibold mb-3">TABLA DE MATERIALES</h1>
             <div className="flex justify-between items-center">
               {roleInSession && roleInSession.name !== "client" && (
-              <Link
-              
-              onClick={() => {
-                localStorage.setItem("type", "Material");
-                toggleResourcesCreate();
-              }}
-              className="mt-1 bg-blue-500 text-white px-4 py-2 mr-5 rounded mb-4 inline-block"
-            >
-              Crear Material
-            </Link>
+                <Link
+                  onClick={() => {
+                    localStorage.setItem("type", "Material");
+                    toggleResourcesCreate();
+                  }}
+                  className="mt-1 bg-blue-500 text-white px-4 py-2 mr-5 rounded mb-4 inline-block"
+                >
+                  Crear Material
+                </Link>
               )}
 
               <input
@@ -248,7 +246,7 @@ const ResourceIndex = () => {
                   </select> */}
                     </th>
 
-                    {roleInSession && roleInSession.name !== "client" &&(
+                    {roleInSession && roleInSession.name !== "client" && (
                       <th className="px-4 py-2">Acciones</th>
                     )}
                   </tr>
@@ -272,12 +270,30 @@ const ResourceIndex = () => {
                           </Link>
 
                           <button
-                            onClick={async () =>
-                              await deleteHandler(resource.id)
-                            }
+                            onClick={async () => {
+                              const result = await Swal.fire({
+                                title: "¿Estás seguro de eliminar el recurso?",
+                                text: "¡No podrás revertir esto!",
+                                icon: "warning",
+                                showCancelButton: true,
+                                confirmButtonColor: "#3085d6",
+                                cancelButtonColor: "#d33",
+                                confirmButtonText: "Sí, eliminarlo",
+                                cancelButtonText: "Cancelar",
+                              });
+
+                              if (result.isConfirmed) {
+                                await deleteHandler(resource.id);
+                                Swal.fire(
+                                  "¡Eliminado!",
+                                  "Tu recurso ha sido eliminado.",
+                                  "success"
+                                );
+                              }
+                            }}
                             className="inline-block bg-red-500 text-white px-4 py-2 rounded"
                           >
-                            Eliminar M
+                            Eliminar
                           </button>
                         </td>
                       )}
@@ -312,18 +328,17 @@ const ResourceIndex = () => {
               TABLA DE PERSONAL
             </h1>
             <div className="flex justify-between items-center">
-            {roleInSession && roleInSession.name !== "client" && (
-              <Link
-               
-               onClick={() => {
-                localStorage.setItem("type", "Personal");
-                toggleResourcesCreate();
-              }}
-                className="bg-blue-500 text-white px-4 py-2 rounded mb-4 inline-block"
-              >
-                Crear Personal
-              </Link>
-            )}
+              {roleInSession && roleInSession.name !== "client" && (
+                <Link
+                  onClick={() => {
+                    localStorage.setItem("type", "Personal");
+                    toggleResourcesCreate();
+                  }}
+                  className="bg-blue-500 text-white px-4 py-2 rounded mb-4 inline-block"
+                >
+                  Crear Personal
+                </Link>
+              )}
               <input
                 type="text"
                 placeholder="Buscar Personal"
@@ -364,7 +379,27 @@ const ResourceIndex = () => {
                           Editar
                         </Link>
                         <button
-                          onClick={() => deleteHandler(resource.id)}
+                          onClick={async () => {
+                            const result = await Swal.fire({
+                              title: "¿Estás seguro de eliminar el personal?",
+                              text: "¡No podrás revertir esto!",
+                              icon: "warning",
+                              showCancelButton: true,
+                              confirmButtonColor: "#3085d6",
+                              cancelButtonColor: "#d33",
+                              confirmButtonText: "Sí, eliminarlo",
+                              cancelButtonText: "Cancelar",
+                            });
+
+                            if (result.isConfirmed) {
+                              await deleteHandler(resource.id);
+                              Swal.fire(
+                                "¡Eliminado!",
+                                "Tu personal ha sido eliminado.",
+                                "success"
+                              );
+                            }
+                          }}
                           className="inline-block bg-red-500 text-white px-4 py-2 rounded"
                         >
                           Eliminar
